Implementação de Vendas.

+++ App/src/main/java/com/UpFest/App/services/venda/PagamentoService.java
@@ -5,6 +5,6 @@ import com.UpFest.App.entities.Pagamento;
 import java.util.List;
 
 public interface PagamentoService {
-    Pagamento validarPagamento(Long entidade, Long referencia, Long valor) throws Exception;
+    Pagamento validarPagamento(int entidade, int referencia, double valor) throws Exception;
     List<Pagamento> getPagamentoToDB(String email) throws Exception;
 }
